fix(browser): require exact Playwright target identity
