--- extensions/browser/src/browser/pw-session.get-page-for-targetid.extension-fallback.test.ts
@@ -1,299 +1,0 @@
-// Browser tests cover pw session.get page for targetid.extension fallback plugin behavior.
-import type { SsrFPolicy } from "openclaw/plugin-sdk/ssrf-runtime";
-import { chromium } from "playwright-core";
-import { afterEach, describe, expect, it, vi } from "vitest";
-import * as chromeModule from "./chrome.js";
-import {
-  closePlaywrightBrowserConnection,
-  getPageForTargetId,
-  listPagesViaPlaywright,
-  setCdpConnectRetryDelayMsForTests,
-} from "./pw-session.js";
-
-const fetchWithSsrFGuardSpy = vi.hoisted(() => vi.fn());
-
-vi.mock("openclaw/plugin-sdk/ssrf-runtime", async (importOriginal) => {
-  const actual = await importOriginal<typeof import("openclaw/plugin-sdk/ssrf-runtime")>();
-  return {
-    ...actual,
-    fetchWithSsrFGuard: (...args: Parameters<typeof actual.fetchWithSsrFGuard>) => {
-      fetchWithSsrFGuardSpy(...args);
-      return actual.fetchWithSsrFGuard(...args);
-    },
-  };
-});
-
-const connectOverCdpSpy = vi.spyOn(chromium, "connectOverCDP");
-const getChromeWebSocketUrlSpy = vi.spyOn(chromeModule, "getChromeWebSocketUrl");
-
-type MockPageSpec = {
-  targetId?: string;
-  url?: string;
-  title?: string;
-};
-
-type BrowserMockBundle = {
-  browser: import("playwright-core").Browser;
-  browserClose: ReturnType<typeof vi.fn>;
-  pages: import("playwright-core").Page[];
-};
-
-type FetchInitWithDispatcher = RequestInit & { dispatcher?: unknown };
-
-function requireFetchCall(fetchSpy: {
-  mock: { calls: Parameters<typeof fetch>[] };
-}): Parameters<typeof fetch> {
-  const [call] = fetchSpy.mock.calls;
-  if (!call) {
-    throw new Error("expected fallback fetch call");
-  }
-  return call;
-}
-
-function requireFetchInit(init: Parameters<typeof fetch>[1]): FetchInitWithDispatcher {
-  if (!init || typeof init !== "object") {
-    throw new Error("expected fallback fetch init");
-  }
-  return init as FetchInitWithDispatcher;
-}
-
-function makeBrowser(pages: MockPageSpec[]): BrowserMockBundle {
-  const browserClose = vi.fn(async () => {});
-  const targetIdByPage = new Map<import("playwright-core").Page, string | undefined>();
-
-  const pageObjects = pages.map((spec, index) => {
-    const page = {
-      on: vi.fn(),
-      context: () => context,
-      title: vi.fn(async () => spec.title ?? spec.targetId ?? `page-${index + 1}`),
-      url: vi.fn(() => spec.url ?? `https://page-${index + 1}.example`),
-    } as unknown as import("playwright-core").Page;
-    targetIdByPage.set(page, spec.targetId);
-    return page;
-  });
-
-  const context: import("playwright-core").BrowserContext = {
-    pages: () => pageObjects,
-    on: vi.fn(),
-    newCDPSession: vi.fn(async (page: import("playwright-core").Page) => ({
-      send: vi.fn(async (method: string) =>
-        method === "Target.getTargetInfo"
-          ? { targetInfo: { targetId: targetIdByPage.get(page) } }
-          : {},
-      ),
-      detach: vi.fn(async () => {}),
-    })),
-  } as unknown as import("playwright-core").BrowserContext;
-
-  const browser = {
-    contexts: () => [context],
-    on: vi.fn(),
-    off: vi.fn(),
-    close: browserClose,
-  } as unknown as import("playwright-core").Browser;
-
-  return { browser, browserClose, pages: pageObjects };
-}
-
-afterEach(async () => {
-  connectOverCdpSpy.mockReset();
-  getChromeWebSocketUrlSpy.mockReset();
-  fetchWithSsrFGuardSpy.mockClear();
-  setCdpConnectRetryDelayMsForTests();
-  await closePlaywrightBrowserConnection().catch(() => {});
-});
-
-function createExtensionFallbackBrowserHarness(options?: {
-  urls?: string[];
-  newCDPSessionError?: string;
-}) {
-  const pageOn = vi.fn();
-  const contextOn = vi.fn();
-  const browserOn = vi.fn();
-  const browserClose = vi.fn(async () => {});
-  const newCDPSession = vi.fn(async () => {
-    throw new Error(options?.newCDPSessionError ?? "Not allowed");
-  });
-
-  const context = {
-    pages: () => [],
-    on: contextOn,
-    newCDPSession,
-  } as unknown as import("playwright-core").BrowserContext;
-
-  const pages = (options?.urls ?? [undefined]).map(
-    (url) =>
-      Object.assign(
-        { on: pageOn, context: () => context },
-        url ? { url: () => url } : {},
-      ) as unknown as import("playwright-core").Page,
-  );
-  (context as unknown as { pages: () => unknown[] }).pages = () => pages;
-
-  const browser = {
-    contexts: () => [context],
-    on: browserOn,
-    close: browserClose,
-  } as unknown as import("playwright-core").Browser;
-
-  connectOverCdpSpy.mockResolvedValue(browser);
-  getChromeWebSocketUrlSpy.mockResolvedValue(null);
-  return { browserClose, newCDPSession, pages };
-}
-
-describe("pw-session getPageForTargetId", () => {
-  it("falls back to the only page when Playwright cannot resolve target ids", async () => {
-    const { browserClose, pages } = createExtensionFallbackBrowserHarness();
-    const [page] = pages;
-
-    const resolved = await getPageForTargetId({
-      cdpUrl: "http://127.0.0.1:18792",
-      targetId: "NOT_A_TAB",
-    });
-    expect(resolved).toBe(page);
-
-    await closePlaywrightBrowserConnection();
-    expect(browserClose).toHaveBeenCalled();
-  });
-
-  it("uses the shared HTTP-base normalization when falling back to /json/list for direct WebSocket CDP URLs", async () => {
-    const [, pageB] = createExtensionFallbackBrowserHarness({
-      urls: ["https://alpha.example", "https://beta.example"],
-    }).pages;
-
-    const fetchSpy = vi.spyOn(globalThis, "fetch").mockResolvedValue(
-      new Response(
-        JSON.stringify([
-          { id: "TARGET_A", url: "https://alpha.example" },
-          { id: "TARGET_B", url: "https://beta.example" },
-        ]),
-        { headers: { "content-type": "application/json" } },
-      ),
-    );
-
-    try {
-      const resolved = await getPageForTargetId({
-        cdpUrl: "ws://127.0.0.1:18792/devtools/browser/SESSION?token=abc",
-        targetId: "TARGET_B",
-      });
-      expect(resolved).toBe(pageB);
-      expect(fetchSpy).toHaveBeenCalledTimes(1);
-      const [fetchUrl, fetchInitOptions] = requireFetchCall(fetchSpy);
-      expect(fetchUrl).toBe("http://127.0.0.1:18792/json/list?token=abc");
-      const fetchInit = requireFetchInit(fetchInitOptions);
-      expect(fetchInit.headers).toEqual({});
-      expect(fetchInit.redirect).toBe("manual");
-      expect(fetchInit.signal).toBeInstanceOf(AbortSignal);
-      expect(fetchInit.dispatcher).toBeUndefined();
-    } finally {
-      fetchSpy.mockRestore();
-    }
-  });
-
-  it("resolves pages from /json/list when page CDP probing fails", async () => {
-    const { newCDPSession, pages } = createExtensionFallbackBrowserHarness({
-      urls: ["https://alpha.example", "https://beta.example"],
-      newCDPSessionError: "Target.attachToBrowserTarget: Not allowed",
-    });
-    const [, pageB] = pages;
-
-    const fetchSpy = vi.spyOn(globalThis, "fetch").mockResolvedValue(
-      new Response(
-        JSON.stringify([
-          { id: "TARGET_A", url: "https://alpha.example" },
-          { id: "TARGET_B", url: "https://beta.example" },
-        ]),
-        { headers: { "content-type": "application/json" } },
-      ),
-    );
-
-    try {
-      const resolved = await getPageForTargetId({
-        cdpUrl: "http://127.0.0.1:19993",
-        targetId: "TARGET_B",
-        ssrfPolicy: {
-          dangerouslyAllowPrivateNetwork: false,
-          allowRfc2544BenchmarkRange: true,
-        },
-      });
-      expect(resolved).toBe(pageB);
-      expect(newCDPSession).toHaveBeenCalled();
-      expect(fetchWithSsrFGuardSpy).toHaveBeenCalledTimes(1);
-      const policy = fetchWithSsrFGuardSpy.mock.calls[0]?.[0]?.policy as SsrFPolicy | undefined;
-      expect(policy?.allowRfc2544BenchmarkRange).toBe(true);
-      expect(policy?.hostnameAllowlist).toEqual(["127.0.0.1"]);
-    } finally {
-      fetchSpy.mockRestore();
-    }
-  });
-
-  it("evicts a stale cached page-less browser once and succeeds on a fresh reconnect", async () => {
-    const stale = makeBrowser([]);
-    const fresh = makeBrowser([{ targetId: "TARGET_OK", url: "https://fresh.example" }]);
-
-    connectOverCdpSpy.mockResolvedValueOnce(stale.browser).mockResolvedValueOnce(fresh.browser);
-    getChromeWebSocketUrlSpy.mockResolvedValue(null);
-
-    await listPagesViaPlaywright({ cdpUrl: "http://127.0.0.1:9222" });
-
-    const resolved = await getPageForTargetId({ cdpUrl: "http://127.0.0.1:9222" });
-
-    expect(resolved).toBe(fresh.pages[0]);
-    expect(connectOverCdpSpy).toHaveBeenCalledTimes(2);
-    expect(stale.browserClose).toHaveBeenCalledTimes(1);
-  });
-
-  it("evicts a stale cached tab-selection miss once and succeeds on a fresh reconnect", async () => {
-    const stale = makeBrowser([
-      { targetId: "TARGET_A", url: "https://alpha.example" },
-      { targetId: "TARGET_C", url: "https://charlie.example" },
-    ]);
-    const fresh = makeBrowser([
-      { targetId: "TARGET_A", url: "https://alpha.example" },
-      { targetId: "TARGET_B", url: "https://beta.example" },
-    ]);
-
-    connectOverCdpSpy.mockResolvedValueOnce(stale.browser).mockResolvedValueOnce(fresh.browser);
-    getChromeWebSocketUrlSpy.mockResolvedValue(null);
-
-    await getPageForTargetId({ cdpUrl: "http://127.0.0.1:9333" });
-
-    const resolved = await getPageForTargetId({
-      cdpUrl: "http://127.0.0.1:9333",
-      targetId: "TARGET_B",
-    });
-
-    expect(resolved).toBe(fresh.pages[1]);
-    expect(connectOverCdpSpy).toHaveBeenCalledTimes(2);
-    expect(stale.browserClose).toHaveBeenCalledTimes(1);
-  });
-
-  it("fails after a single reconnect when the refreshed browser is still page-less", async () => {
-    const stale = makeBrowser([]);
-    const stillBroken = makeBrowser([]);
-
-    connectOverCdpSpy
-      .mockResolvedValueOnce(stale.browser)
-      .mockResolvedValueOnce(stillBroken.browser);
-    getChromeWebSocketUrlSpy.mockResolvedValue(null);
-
-    await listPagesViaPlaywright({ cdpUrl: "http://127.0.0.1:9444" });
-
-    await expect(getPageForTargetId({ cdpUrl: "http://127.0.0.1:9444" })).rejects.toThrow(
-      "No pages available in the connected browser.",
-    );
-    expect(connectOverCdpSpy).toHaveBeenCalledTimes(2);
-    expect(stale.browserClose).toHaveBeenCalledTimes(1);
-  });
-
-  it("does not add an extra top-level retry for non-recoverable connect failures", async () => {
-    setCdpConnectRetryDelayMsForTests(0);
-    connectOverCdpSpy.mockRejectedValue(new Error("connectOverCDP exploded"));
-    getChromeWebSocketUrlSpy.mockResolvedValue(null);
-
-    await expect(getPageForTargetId({ cdpUrl: "http://127.0.0.1:9555" })).rejects.toThrow(
-      "connectOverCDP exploded",
-    );
-    expect(connectOverCdpSpy).toHaveBeenCalledTimes(3);
-  });
-});
